+++ TinkoffASDKUI/TinkoffASDKUI/Cards/PaymentCardPredicate.swift
@@ -1,6 +1,6 @@
 //
 //
-//  CardsController.swift
+//  PaymentCardPredicate.swift
 //
 //  Copyright (c) 2021 Tinkoff Bank
 //
@@ -18,18 +18,21 @@
 //
 
 
-import Foundation
+import TinkoffASDKCore
 
-protocol CardsControllerListener: AnyObject {
-    func cardsControllerDidUpdateCards(_ cardsController: CardsController)
-}
-
-protocol CardsController {
-    var customerKey: String { get }
+public struct PaymentCardPredicate {
     
-    func loadCards(completion: @escaping (Result<[PaymentCard], Error>) -> Void)
-    func getCards(predicates: PaymentCardPredicate...) -> [PaymentCard]
+    static var activeCards: PaymentCardPredicate {
+        PaymentCardPredicate(closure: { $0.status == .active })
+    }
     
-    func addListener(_ listener: CardsControllerListener)
-    func removeListener(_ listener: CardsControllerListener)
+    static var parentPaymentCards: PaymentCardPredicate {
+        PaymentCardPredicate(closure: { $0.parentPaymentId != nil })
+    }
+    
+    let closure: (PaymentCard) -> Bool
+    
+    init(closure: @escaping (PaymentCard) -> Bool) {
+        self.closure = closure
+    }
 }
